+++ src/main/java/com/mamoori/mamooriback/api/controller/CommunityController.java
@@ -2,7 +2,7 @@ package com.mamoori.mamooriback.api.controller;
 
 import com.mamoori.mamooriback.api.dto.PostResponse;
 import com.mamoori.mamooriback.api.service.PostService;
-import com.mamoori.mamooriback.api.controller.request.PostRequest;
+import com.mamoori.mamooriback.api.dto.PostRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
