+++ src/main/java/com/mamoori/mamooriback/api/repository/UserRepository.java
@@ -6,6 +6,6 @@ import com.mamoori.mamooriback.api.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-	Optional<User> findByEmail(String email); //
+	Optional<User> findByEmail(String email);
 	boolean existsByEmail(String email);
 }
