+++ src/main/java/com/mamoori/mamooriback/api/service/PostService.java
@@ -1,6 +1,6 @@
 package com.mamoori.mamooriback.api.service;
 
-import com.mamoori.mamooriback.api.controller.request.PostRequest;
+import com.mamoori.mamooriback.api.dto.PostRequest;
 import com.mamoori.mamooriback.api.dto.PostResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
